remove research related code from main code path
